wordpress_scraping_pipeline/backend/requirements.txt
# WordPress Scraping Pipeline Dependencies
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-cloud-bigquery==3.13.0
python-dotenv==1.0.0
tqdm==4.66.1
supabase>=2.0.0
